Handle <br>s inside code blocks.
Context: I'm processing some exported code which has code that looks like:

<pre><code>foo<br>bar<br>baz</code></pre>

And turndown is not accounting for those line breaks.

We could use innerText instead of textContent to account for them, but it's not
supported in JSDOM (https://github.com/jsdom/jsdom/issues/1245).

We could implement a crappy version of it which includes newlines generated by
<br>, but it seems better to teach the newline replacement about preformatted
text, and just use the already-processed content instead.